# docs/swagger.yaml
basePath: /api
definitions:
...
  api.EmptyResponse:
    properties:
      data: {}
      meta: {}
    type: object
  api.ErrorDetail:
    properties:
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorDetail'
    type: object
...
  api.PaginationMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  api.TeamListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Team'
        type: array
      meta:
        $ref: '#/definitions/api.PaginationMeta'
    type: object
  api.TeamProfileListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.TeamProfile'
        type: array
      meta:
        $ref: '#/definitions/api.PaginationMeta'
    type: object
  api.TeamProfileResponse:
    properties:
      data:
        $ref: '#/definitions/models.TeamProfile'
      meta: {}
    type: object
  api.TeamResponse:
    properties:
      data:
        $ref: '#/definitions/models.Team'
      meta: {}
    type: object
...
  api.teamProfileRequest:
    properties:
      coach_name:
        type: string
      contact_email:
        type: string
      headquarters:
        type: string
      sponsor_info:
        type: string
      team_id:
        type: integer
      website:
        type: string
    required:
    - team_id
    type: object
  api.teamRequest:
    properties:
      country_code:
        type: string
      discipline_id:
        type: integer
      is_verified:
        type: boolean
      logo_url:
        type: string
      name:
        type: string
      tag:
        type: string
      world_ranking:
        type: number
    required:
    - country_code
    - discipline_id
    - name
    - tag
    type: object
...
  models.Team:
    properties:
      country_code:
        type: string
      created_at:
        type: string
      discipline_id:
        type: integer
      id:
        type: integer
      is_verified:
        type: boolean
      logo_url:
        type: string
      name:
        type: string
      tag:
        type: string
      world_ranking:
        type: number
    type: object
  models.TeamProfile:
    properties:
      coach_name:
        type: string
      contact_email:
        type: string
      headquarters:
        type: string
      sponsor_info:
        type: string
      team_id:
        type: integer
      website:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for disciplines, teams, tournaments, matches, and reports.
  title: DB Course Project API
  version: "1.0"
paths:
...
  /team-profiles:
    get:
      parameters:
      - description: Team ID
        in: query
        name: team_id
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TeamProfileListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List team profiles
      tags:
      - TeamProfiles
    post:
      consumes:
      - application/json
      parameters:
      - description: Team profile payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.teamProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.TeamProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create team profile
      tags:
      - TeamProfiles
  /team-profiles/{team_id}:
    delete:
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/api.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete team profile
      tags:
      - TeamProfiles
    get:
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TeamProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get team profile
      tags:
      - TeamProfiles
    put:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Team profile payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.teamProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TeamProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update team profile
      tags:
      - TeamProfiles
  /teams:
    get:
      parameters:
      - description: Search by name or tag
        in: query
        name: search
        type: string
      - description: Country code
        in: query
        name: country_code
        type: string
      - description: Discipline ID
        in: query
        name: discipline_id
        type: integer
      - description: Verification flag
        in: query
        name: is_verified
        type: boolean
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TeamListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List teams
      tags:
      - Teams
    post:
      consumes:
      - application/json
      parameters:
      - description: Team payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.teamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create team
      tags:
      - Teams
  /teams/{id}:
    delete:
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/api.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete team
      tags:
      - Teams
    get:
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get team
      tags:
      - Teams
    put:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: Team payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.teamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update team
      tags:
      - Teams
...
swagger: "2.0"